# Conflicts:
#	src/test/scala/b4processor/modules/memory/DataMemoryTest.scala

+++ src/test/scala/b4processor/utils/ALUValue.scala
@@ -1,3 +1,3 @@
 package b4processor.utils
 
-case class ALUValue(destinationTag: Int, value: Int)
+case class ALUValue(destinationTag: Int, value: BigInt)
